[TIM-911] Make output more flexible.
